refactor: export help fns to helpers file
